docs: change markdown headings
At the top of README.md, there was a line goes like this: `#vim-header`;
using the `===` Markdown syntax to make sure that this line will be
rendered as a H1 heading.

The similar change applies for `###Thanks to Contributors` line,
changing it to be rendered as a H3 heading.

This could be a matter of style preference but all other H1 headings
have been converted to H2 headings.