Modify headers' update behavior
This patch modify existing behavior in order to update headers.

The plugin now loops through global options to see which ones are set in
$MYVIMRC from which a list is built. Then, it iterates over the list to
see if all headers are present from the start of the file to a certain
threshold (option g:header_max_size, default is 5).

If one required header is missing, it considers that headers aren't set
so it adds new ones. Otherwise, it updates headers keeping positions and
identations.

Patch also updates the README.md and CHANGELOG.md to reflect thoses
changes.